feat: add new route to execute job of deployment status analytics + fix mongo session leak
