+++ server/server.go
@@ -9,7 +9,7 @@ import (
 	"github.com/soprasteria/dad/server/auth"
 	"github.com/soprasteria/dad/server/controllers"
 	"github.com/soprasteria/dad/server/email"
-	"github.com/soprasteria/dad/server/schedule"
+	"github.com/soprasteria/dad/server/jobs"
 	"github.com/soprasteria/dad/server/types"
 	"github.com/spf13/viper"
 )
@@ -28,6 +28,7 @@ func New(version string) {
 	projectsC := controllers.Projects{}
 	technologiesC := controllers.Technologies{}
 	exportC := controllers.Export{}
+	adminC := controllers.Admin{}
 
 	engine.Use(middleware.Logger())
 	engine.Use(middleware.Recover())
@@ -132,6 +133,13 @@ func New(version string) {
 			exportAPI.Use(getAuthenticatedUser)
 			exportAPI.GET("", exportC.ExportAll)
 		}
+
+		adminAPI := api.Group("/admin")
+		{
+			adminAPI.Use(hasRole(types.AdminRole))
+			jobsAPI := adminAPI.Group("/jobs")
+			jobsAPI.POST("/deployment-indicators", adminC.ExecuteDeploymentJobAnalytics)
+		}
 	}
 
 	engine.Static("/js", "client/js")
@@ -148,7 +156,7 @@ func New(version string) {
 	}
 
 	// Launch back-end tasks.
-	go schedule.RunBackgroundJobs()
+	go jobs.RunBackgroundJobs()
 
 	if err := engine.Start(":8080"); err != nil {
 		engine.Logger.Fatal(err.Error())
